relay_event_schema::processor
Function process_chunked_value

pub fn process_chunked_value<F>(value: &mut String, meta: &mut Meta, f: F)
where
    F: FnOnce(Vec<Chunk<'_>>) -> Vec<Chunk<'_>>,

Splits the string into chunks, maps each chunk and then joins chunks again, emitting remarks along the process.