relay_event_schema::protocol

Enum SpanStatus

#[repr(u8)]
pub enum SpanStatus {

    Ok = 0,
    Cancelled = 1,
    Unknown = 2,
    InvalidArgument = 3,
    DeadlineExceeded = 4,
    NotFound = 5,
    AlreadyExists = 6,
    PermissionDenied = 7,
    ResourceExhausted = 8,
    FailedPrecondition = 9,
    Aborted = 10,
    OutOfRange = 11,
    Unimplemented = 12,
    InternalError = 13,
    Unavailable = 14,
    DataLoss = 15,
    Unauthenticated = 16,

}

Trace status.

Values from https://github.com/open-telemetry/opentelemetry-specification/blob/8fb6c14e4709e75a9aaa64b0dbbdf02a6067682a/specification/api-tracing.md#status Mapping to HTTP from https://github.com/open-telemetry/opentelemetry-specification/blob/8fb6c14e4709e75a9aaa64b0dbbdf02a6067682a/specification/data-http.md#status

Variants

Ok = 0

The operation completed successfully.

HTTP status 100..299 + successful redirects from the 3xx range.

Cancelled = 1

The operation was cancelled (typically by the user).

Unknown = 2

Unknown. Any non-standard HTTP status code.

โWe do not know whether the transaction failed or succeededโ

InvalidArgument = 3

Client specified an invalid argument. 4xx.

Note that this differs from FailedPrecondition. InvalidArgument indicates arguments that are problematic regardless of the state of the system.

DeadlineExceeded = 4

Deadline expired before operation could complete.

For operations that change the state of the system, this error may be returned even if the operation has been completed successfully.

HTTP redirect loops and 504 Gateway Timeout

NotFound = 5

404 Not Found. Some requested entity (file or directory) was not found.

AlreadyExists = 6

Already exists (409)

Some entity that we attempted to create already exists.

PermissionDenied = 7

403 Forbidden

The caller does not have permission to execute the specified operation.

ResourceExhausted = 8

429 Too Many Requests

Some resource has been exhausted, perhaps a per-user quota or perhaps the entire file system is out of space.

FailedPrecondition = 9

Operation was rejected because the system is not in a state required for the operationโs execution

Aborted = 10

The operation was aborted, typically due to a concurrency issue.

OutOfRange = 11

Operation was attempted past the valid range.

Unimplemented = 12

501 Not Implemented

Operation is not implemented or not enabled.

InternalError = 13

Other/generic 5xx.

Unavailable = 14

503 Service Unavailable

DataLoss = 15

Unrecoverable data loss or corruption

Unauthenticated = 16

401 Unauthorized (actually does mean unauthenticated according to RFC 7235)

Prefer PermissionDenied if a user is logged in.

Implementations

impl SpanStatus

pub fn as_str(&self) -> &'static str

Returns the string representation of the status.