relay_event_schema::protocol

Enum TransactionSource

pub enum TransactionSource {
    Custom,
    Url,
    Route,
    View,
    Component,
    Sanitized,
    Task,
    Unknown,
    Other(String),
}

Describes how the name of the transaction was determined.

Variants

Custom

User-defined name set through set_transaction_name.

Url

Raw URL, potentially containing identifiers.

Route

Parametrized URL or route.

View

Name of the view handling the request.

Component

Named after a software component, such as a function or class name.

Sanitized

The transaction name was updated to reduce the name cardinality.

Task

Name of a background task (e.g. a Celery task).

Unknown

This is the default value set by Relay for legacy SDKs.

Other(String)

Any other unknown source that is not explicitly defined above.

Implementations

impl TransactionSource

pub fn as_str(&self) -> &str