relay_event_schema::protocol

Function attribute_pii_from_conventions

pub fn attribute_pii_from_conventions(state: &ProcessingState<'_>) -> Pii

Determines the Pii value for an attribute (or, more exactly, the attribute’s value field) by looking it up in relay-conventions.

The attribute’s value may be addressed by <key>.value (for advanced scrubbing rules) or by just <key> (for default scrubbing rules). Therefore, we iterate backwards through the processing state’s segments and try to look up each in the conventions.

If the attribute is not found in the conventions, this returns Pii::True as a precaution.