relay_event_schema::protocol

Module process_state

Nested message and enum types in ProcessState.

Modules

file_system_state

Nested message and enum types in FileSystemState.

memory_state

Nested message and enum types in MemoryState.

Structs

FileSystemStateContext

Records the state of the file system at the time of crash.

MemoryStateContext

Records the state of process memory at the time of crash. Next id: 6