relay_event_schema::protocol
Struct SentryTags

pub struct SentryTags {

    pub release: Annotated<String>,
    pub user: Annotated<String>,
    pub user_id: Annotated<String>,
    pub user_ip: Annotated<String>,
    pub user_username: Annotated<String>,
    pub user_email: Annotated<String>,
    pub environment: Annotated<String>,
    pub transaction: Annotated<String>,
    pub transaction_method: Annotated<String>,
    pub transaction_op: Annotated<String>,
    pub browser_name: Annotated<String>,
    pub sdk_name: Annotated<String>,
    pub sdk_version: Annotated<String>,
    pub platform: Annotated<String>,
    pub mobile: Annotated<String>,
    pub device_class: Annotated<String>,
    pub os_name: Annotated<String>,
    pub action: Annotated<String>,
    pub ai_pipeline_group: Annotated<String>,
    pub category: Annotated<String>,
    pub description: Annotated<String>,
    pub domain: Annotated<String>,
    pub raw_domain: Annotated<String>,
    pub group: Annotated<String>,
    pub http_decoded_response_content_length: Annotated<String>,
    pub http_response_content_length: Annotated<String>,
    pub http_response_transfer_size: Annotated<String>,
    pub resource_render_blocking_status: Annotated<String>,
    pub op: Annotated<String>,
    pub status: Annotated<String>,
    pub status_code: Annotated<String>,
    pub system: Annotated<String>,
    pub ttid: Annotated<String>,
    pub ttfd: Annotated<String>,
    pub file_extension: Annotated<String>,
    pub main_thread: Annotated<String>,
    pub app_start_type: Annotated<String>,
    pub replay_id: Annotated<String>,
    pub cache_hit: Annotated<String>,
    pub cache_key: Annotated<String>,
    pub trace_status: Annotated<String>,
    pub messaging_destination_name: Annotated<String>,
    pub messaging_message_id: Annotated<String>,
    pub thread_name: Annotated<String>,
    pub thread_id: Annotated<String>,
    pub profiler_id: Annotated<String>,
    pub user_country_code: Annotated<String>,
    pub user_subregion: Annotated<String>,

}

Indexable fields added by sentry (server-side).

Fields
release: Annotated<String>user: Annotated<String>user_id: Annotated<String>user_ip: Annotated<String>user_username: Annotated<String>user_email: Annotated<String>environment: Annotated<String>transaction: Annotated<String>transaction_method: Annotated<String>transaction_op: Annotated<String>browser_name: Annotated<String>sdk_name: Annotated<String>sdk_version: Annotated<String>platform: Annotated<String>mobile: Annotated<String>device_class: Annotated<String>os_name: Annotated<String>action: Annotated<String>ai_pipeline_group: Annotated<String>

The group of the ancestral span with op ai.pipeline.(String)*

category: Annotated<String>description: Annotated<String>domain: Annotated<String>raw_domain: Annotated<String>group: Annotated<String>http_decoded_response_content_length: Annotated<String>http_response_content_length: Annotated<String>http_response_transfer_size: Annotated<String>resource_render_blocking_status: Annotated<String>op: Annotated<String>status: Annotated<String>status_code: Annotated<String>system: Annotated<String>ttid: Annotated<String>

Contributes to Time-To-Initial-Display(String).

ttfd: Annotated<String>

Contributes to Time-To-Full-Display(String).

file_extension: Annotated<String>

File extension for resource spans(String).

main_thread: Annotated<String>

Span started on main thread(String).

app_start_type: Annotated<String>

The start type of the application when the span occurred(String).

replay_id: Annotated<String>cache_hit: Annotated<String>cache_key: Annotated<String>trace_status: Annotated<String>messaging_destination_name: Annotated<String>messaging_message_id: Annotated<String>thread_name: Annotated<String>thread_id: Annotated<String>profiler_id: Annotated<String>user_country_code: Annotated<String>user_subregion: Annotated<String>