relay_event_schema::protocol

Struct Span

pub struct Span {

    pub timestamp: Annotated<Timestamp>,
    pub start_timestamp: Annotated<Timestamp>,
    pub exclusive_time: Annotated<f64>,
    pub op: Annotated<OperationType>,
    pub span_id: Annotated<SpanId>,
    pub parent_span_id: Annotated<SpanId>,
    pub trace_id: Annotated<TraceId>,
    pub segment_id: Annotated<SpanId>,
    pub is_segment: Annotated<bool>,
    pub is_remote: Annotated<bool>,
    pub status: Annotated<SpanStatus>,
    pub description: Annotated<String>,
    pub tags: Annotated<Object<JsonLenientString>>,
    pub origin: Annotated<OriginType>,
    pub profile_id: Annotated<EventId>,
    pub data: Annotated<SpanData>,
    pub links: Annotated<Array<SpanLink>>,
    pub sentry_tags: Annotated<SentryTags>,
    pub received: Annotated<Timestamp>,
    pub measurements: Annotated<Measurements>,
    pub platform: Annotated<String>,
    pub was_transaction: Annotated<bool>,
    pub kind: Annotated<SpanKind>,
    pub performance_issues_spans: Annotated<bool>,
    pub other: Object<Value>,

}

Fields
timestamp: Annotated<Timestamp>

Timestamp when the span was ended.

start_timestamp: Annotated<Timestamp>

Timestamp when the span started.

exclusive_time: Annotated<f64>

The amount of time in milliseconds spent in this span, excluding its immediate child spans.

op: Annotated<OperationType>

Span type (see OperationType docs).

span_id: Annotated<SpanId>

The Span id.

parent_span_id: Annotated<SpanId>

The ID of the span enclosing this span.

trace_id: Annotated<TraceId>

The ID of the trace the span belongs to.

segment_id: Annotated<SpanId>

A unique identifier for a segment within a trace (8 byte hexadecimal string).

For spans embedded in transactions, the segment_id is the span_id of the containing transaction.

is_segment: Annotated<bool>

Whether or not the current span is the root of the segment.

is_remote: Annotated<bool>

Indicates whether a span’s parent is remote.

For OpenTelemetry spans, this is derived from span flags bits 8 and 9. See SPAN_FLAGS_CONTEXT_HAS_IS_REMOTE_MASK and SPAN_FLAGS_CONTEXT_IS_REMOTE_MASK.

The states are:

empty: unknown

false: is not remote

true: is remote

status: Annotated<SpanStatus>

The status of a span.

description: Annotated<String>

Human readable description of a span (e.g. method URL).

tags: Annotated<Object<JsonLenientString>>

Arbitrary tags on a span, like on the top-level event.

origin: Annotated<OriginType>

The origin of the span indicates what created the span (see OriginType docs).

profile_id: Annotated<EventId>

ID of a profile that can be associated with the span.

data: Annotated<SpanData>

Arbitrary additional data on a span.

Besides arbitrary user data, this object also contains SDK-provided fields used by the product (see https://develop.sentry.dev/sdk/performance/span-data-conventions/).

links: Annotated<Array<SpanLink>>

Links from this span to other spans

sentry_tags: Annotated<SentryTags>

Tags generated by Relay. These tags are a superset of the tags set on span metrics.

received: Annotated<Timestamp>

Timestamp when the span has been received by Sentry.

measurements: Annotated<Measurements>

Measurements which holds observed values such as web vitals.

platform: Annotated<String>

Platform identifier.

See Event::platform.

was_transaction: Annotated<bool>

Whether the span is a segment span that was converted from a transaction.

kind: Annotated<SpanKind>performance_issues_spans: Annotated<bool>

Temporary flag that controls where performance issues are detected.

When the flag is set to true, performance issues will be detected on this span provided it is a root (segment) instead of the transaction event.

Only set on root spans extracted from transactions.

other: Object<Value>

Additional arbitrary fields for forwards compatibility.