relay_event_schema::protocol

Trait AsPair

pub trait AsPair {
    type Key: AsRef<str>;
    type Value: ProcessValue;

    // Required methods
    fn from_pair(pair: (Annotated<Self::Key>, Annotated<Self::Value>)) -> Self;
    fn into_pair(self) -> (Annotated<Self::Key>, Annotated<Self::Value>);
    fn as_pair(&self) -> (&Annotated<Self::Key>, &Annotated<Self::Value>);
    fn as_pair_mut(
        &mut self,
    ) -> (&mut Annotated<Self::Key>, &mut Annotated<Self::Value>);

    // Provided methods
    fn key(&self) -> Option<&str> { ... }
    fn value(&self) -> Option<&Self::Value> { ... }
}

A trait to abstract over pairs.

Required Associated Types

type Key: AsRef<str>

type Value: ProcessValue

Required Methods

fn from_pair(pair: (Annotated<Self::Key>, Annotated<Self::Value>)) -> Self

Constructs this value from a raw tuple.

fn into_pair(self) -> (Annotated<Self::Key>, Annotated<Self::Value>)

Converts this pair into a raw tuple.

fn as_pair(&self) -> (&Annotated<Self::Key>, &Annotated<Self::Value>)

Extracts a key and value pair from the object.

fn as_pair_mut( &mut self, ) -> (&mut Annotated<Self::Key>, &mut Annotated<Self::Value>)

Extracts the mutable key and value pair from the object.

Provided Methods

fn key(&self) -> Option<&str>

Returns a reference to the string representation of the key.

fn value(&self) -> Option<&Self::Value>

Returns a reference to the value.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.