relay_event_schema
Module processor

Provides support for processing structures.

Macros

enum_set

Creates a EnumSet literal, which can be used in const contexts.

Structs

CharacterSet

A set of characters allowed or denied for a (string) field.

EnumSet

An efficient set type for enums.

FieldAttrs

Meta information about a field.

Path

Represents the ProcessingState as a path.

ProcessingState

An event’s processing state.

ProcessingStateIter

UnknownValueTypeError

Error for unknown value types.

Enums

Chunk

A type for dealing with chunks of annotated text.

Pii

Whether an attribute should be PII-strippable/should be subject to datascrubbers

ProcessingAction

Used to indicate how to handle an annotated value in a callback.

ValueType

The (simplified) type of a value.

Traits

ProcessValue

A recursively processable value.

Processor

A trait for processing processable values.

Functions

apply

Modifies this value based on the action returned by f.

join_chunks

Concatenates chunks into a string and emits remarks for redacted sections.

process_chunked_value

Splits the string into chunks, maps each chunk and then joins chunks again, emitting remarks along the process.

process_value

Processes the value using the given processor.

split_chunks

Chunks the given text based on remarks.

Type Aliases

ProcessingResult

The result of running a processor on a value implementing ProcessValue.

Derive Macros

ProcessValue

Derive the ProcessValue trait.